Added separate compilation unit for dlmalloc
Made a wrapper that just deals with dlmalloc, and then separated
out the logic for dealing with debug versus non-debug allocator.
